+++ src/main/java/xlxacidxlx/datmod/register/Recipes.java
@@ -14,442 +14,369 @@ public class Recipes {
 		registerSmeltingRecipes();
 	}
 
-	public static void registerRecipe(ItemStack output, String lineOne, String lineTwo, String lineThree, ItemStack input) {
-		GameRegistry.addRecipe(output, lineOne, lineTwo, lineThree, input);
-	}
-
-	public static void registerRecipe(ItemStack output, Object... params) {
-		GameRegistry.addRecipe(output, params);
-	}
-
 	private static void registerRecipes() {
 		Item diamondium = Items.diamondium;
 		Item emeraldi = Items.emeraldi;
 		Item goldiriteIngot = Items.goldiriteIngot;
 		Item ironiumIngot = Items.ironiumIngot;
 
-		ItemStack lapis = new ItemStack(net.minecraft.init.Items.DYE, 1, 2);
+		ItemStack lapis = new ItemStack(net.minecraft.init.Items.DYE, 1, 4);
 		net.minecraft.item.Item stick = net.minecraft.init.Items.STICK;
 		net.minecraft.item.Item redstone = net.minecraft.init.Items.REDSTONE;
 
 		// Axes
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumAxe), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumAxe),
 				" DD",
 				" SD",
 				" S ",
 				'D', diamondium,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiAxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiAxe),
 				" EE",
 				" SE",
 				" S ",
 				'E', emeraldi,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiriteAxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiriteAxe),
 				" GG",
 				" SG",
 				" S ",
 				'G', goldiriteIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumAxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumAxe),
 				" II",
 				" SI",
 				" S ",
 				'I', ironiumIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliAxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliAxe),
 				" LL",
 				" SL",
 				" S ",
 				'L', lapis,
-				'S', stick
-		});
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.redstoneAxe),
+				" RR",
+				" SR",
+				" S ",
+				'R', redstone,
+				'S', stick);
 
 		// Blocks
-		GameRegistry.addRecipe(new ItemStack(Blocks.diamondium), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Blocks.diamondium),
 				"DDD",
 				"DDD",
 				"DDD",
-				'D', diamondium
-		});
-		GameRegistry.addRecipe(new ItemStack(Blocks.emeraldi), new Object[]{
+				'D', diamondium);
+		GameRegistry.addRecipe(new ItemStack(Blocks.emeraldi),
 				"EEE",
 				"EEE",
 				"EEE",
-				'E', emeraldi
-		});
-		GameRegistry.addRecipe(new ItemStack(Blocks.goldirite), new Object[]{
+				'E', emeraldi);
+		GameRegistry.addRecipe(new ItemStack(Blocks.goldirite),
 				"GGG",
 				"GGG",
 				"GGG",
-				'G', goldiriteIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Blocks.ironium), new Object[]{
+				'G', goldiriteIngot);
+		GameRegistry.addRecipe(new ItemStack(Blocks.ironium),
 				"III",
 				"III",
 				"III",
-				'I', ironiumIngot
-		});
+				'I', ironiumIngot);
 
 		// Boots
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumBoots), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumBoots),
 				"D D",
 				"D D",
 				"   ",
-				'D', diamondium
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiBoots), new Object[]{
+				'D', diamondium);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiBoots),
 				"E E",
 				"E E",
 				"   ",
-				'E', emeraldi
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiriteBoots), new Object[]{
+				'E', emeraldi);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiriteBoots),
 				"G G",
 				"G G",
 				"   ",
-				'G', goldiriteIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumBoots), new Object[]{
+				'G', goldiriteIngot);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumBoots),
 				"I I",
 				"I I",
 				"   ",
-				'I', ironiumIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliBoots), new Object[]{
+				'I', ironiumIngot);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliBoots),
 				"L L",
 				"L L",
 				"   ",
-				'L', lapis
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.redstoneBoots), new Object[]{
+				'L', lapis);
+		GameRegistry.addRecipe(new ItemStack(Items.redstoneBoots),
 				"R R",
 				"R R",
 				"   ",
-				'R', redstone
-		});
+				'R', redstone);
 
 		// Chestplates
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumChestplate), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumChestplate),
 				"D D",
 				"DDD",
 				"DDD",
-				'D', diamondium
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiChestplate), new Object[]{
+				'D', diamondium);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiChestplate),
 				"E E",
 				"EEE",
 				"EEE",
-				'E', emeraldi
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiriteChestplate), new Object[]{
+				'E', emeraldi);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiriteChestplate),
 				"G G",
 				"GGG",
 				"GGG",
-				'G', goldiriteIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumChestplate), new Object[]{
+				'G', goldiriteIngot);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumChestplate),
 				"I I",
 				"III",
 				"III",
-				'I', ironiumIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliChestplate), new Object[]{
+				'I', ironiumIngot);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliChestplate),
 				"L L",
 				"LLL",
 				"LLL",
-				'L', lapis
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.redstoneChestplate), new Object[]{
+				'L', lapis);
+		GameRegistry.addRecipe(new ItemStack(Items.redstoneChestplate),
 				"R R",
 				"RRR",
 				"RRR",
-				'R', redstone
-		});
-
-		// Compressed Items
-		// TODO: Figure out how to register the recipes for compressed items..
+				'R', redstone);
 
 		// Helmets
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumHelmet), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumHelmet),
 				"DDD",
 				"D D",
 				"   ",
-				'D', diamondium
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiHelmet), new Object[]{
+				'D', diamondium);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiHelmet),
 				"EEE",
 				"E E",
 				"   ",
-				'E', emeraldi
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiriteHelmet), new Object[]{
+				'E', emeraldi);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiriteHelmet),
 				"GGG",
 				"G G",
 				"   ",
-				'G', goldiriteIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumHelmet), new Object[]{
+				'G', goldiriteIngot);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumHelmet),
 				"III",
 				"I I",
 				"   ",
-				'I', ironiumIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliHelmet), new Object[]{
+				'I', ironiumIngot);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliHelmet),
 				"LLL",
 				"L L",
 				"   ",
-				'L', lapis
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.redstoneHelmet), new Object[]{
+				'L', lapis);
+		GameRegistry.addRecipe(new ItemStack(Items.redstoneHelmet),
 				"RRR",
 				"R R",
 				"   ",
-				'R', redstone
-		});
+				'R', redstone);
 
 		// Hoes
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumHoe), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumHoe),
 				" DD",
 				" S ",
 				" S ",
 				'D', diamondium,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiHoe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiHoe),
 				" EE",
 				" S ",
 				" S ",
 				'E', emeraldi,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiriteHoe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiriteHoe),
 				" GG",
 				" S ",
 				" S ",
 				'G', goldiriteIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumHoe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumHoe),
 				" II",
 				" S ",
 				" S ",
 				'I', ironiumIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliHoe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliHoe),
 				" LL",
 				" S ",
 				" S ",
 				'L', lapis,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.redstoneHoe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.redstoneHoe),
 				" RR",
 				" S ",
 				" S ",
 				'R', redstone,
-				'S', stick
-		});
+				'S', stick);
 
 		// Leggings
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumLeggings), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumLeggings),
 				"DDD",
 				"D D",
 				"D D",
-				'D', diamondium
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiLeggings), new Object[]{
+				'D', diamondium);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiLeggings),
 				"EEE",
 				"E E",
 				"E E",
-				'E', emeraldi
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiriteLeggings), new Object[]{
+				'E', emeraldi);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiriteLeggings),
 				"GGG",
 				"G G",
 				"G G",
-				'G', goldiriteIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumLeggings), new Object[]{
+				'G', goldiriteIngot);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumLeggings),
 				"III",
 				"I I",
 				"I I",
-				'I', ironiumIngot
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliLeggings), new Object[]{
+				'I', ironiumIngot);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliLeggings),
 				"LLL",
 				"L L",
 				"L L",
-				'L', lapis
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.redstoneLeggings), new Object[]{
+				'L', lapis);
+		GameRegistry.addRecipe(new ItemStack(Items.redstoneLeggings),
 				"RRR",
 				"R R",
 				"R R",
-				'R', redstone
-		});
+				'R', redstone);
 
 		// Magnet
-		GameRegistry.addRecipe(new ItemStack(Items.magnetItem), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.magnetItem),
 				"III",
 				"IRI",
 				"I I",
 				'I', ironiumIngot,
-				'R', redstone
-		});
+				'R', redstone);
 
 		// Pickaxes
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumPickaxe), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumPickaxe),
 				"DDD",
 				" S ",
 				" S ",
 				'D', diamondium,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiPickaxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiPickaxe),
 				"EEE",
 				" S ",
 				" S ",
 				'E', emeraldi,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiritePickaxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiritePickaxe),
 				"GGG",
 				" S ",
 				" S ",
 				'G', goldiriteIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumPickaxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumPickaxe),
 				"III",
 				" S ",
 				" S ",
 				'I', ironiumIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliPickaxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliPickaxe),
 				"LLL",
 				" S ",
 				" S ",
 				'L', lapis,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.redstonePickaxe), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.redstonePickaxe),
 				"RRR",
 				" S ",
 				" S ",
 				'R', redstone,
-				'S', stick
-		});
+				'S', stick);
 
 		// Shovels
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumShovel), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumShovel),
 				" D ",
 				" S ",
 				" S ",
 				'D', diamondium,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiShovel), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiShovel),
 				" E ",
 				" S ",
 				" S ",
 				'E', emeraldi,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiriteShovel), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiriteShovel),
 				" G ",
 				" S ",
 				" S ",
 				'G', goldiriteIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumShovel), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumShovel),
 				" I ",
 				" S ",
 				" S ",
 				'I', ironiumIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliShovel), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliShovel),
 				" L ",
 				" S ",
 				" S ",
 				'L', lapis,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.redstoneShovel), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.redstoneShovel),
 				" R ",
 				" S ",
 				" S ",
 				'R', redstone,
-				'S', stick
-		});
+				'S', stick);
 
 		// Swords
-		GameRegistry.addRecipe(new ItemStack(Items.diamondiumSword), new Object[]{
+		GameRegistry.addRecipe(new ItemStack(Items.diamondiumSword),
 				" D ",
 				" D ",
 				" S ",
 				'D', diamondium,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.emeraldiSword), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.emeraldiSword),
 				" E ",
 				" E ",
 				" S ",
 				'E', emeraldi,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.goldiriteSword), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.goldiriteSword),
 				" G ",
 				" G ",
 				" S ",
 				'G', goldiriteIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.ironiumSword), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.ironiumSword),
 				" I ",
 				" I ",
 				" S ",
 				'I', ironiumIngot,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliSword), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.lapisLazuliSword),
 				" L ",
 				" L ",
 				" S ",
 				'L', lapis,
-				'S', stick
-		});
-		GameRegistry.addRecipe(new ItemStack(Items.redstoneSword), new Object[]{
+				'S', stick);
+		GameRegistry.addRecipe(new ItemStack(Items.redstoneSword),
 				" R ",
 				" R ",
 				" S ",
 				'R', redstone,
-				'S', stick
-		});
+				'S', stick);
 	}
 
-	/**
-	 * Registers a smelting recipe
-	 *
-	 * @param input  The item to put in to smelt
-	 * @param output The item received from smelting
-	 * @param xp     The amount of XP gained from smelting the item
-	 */
 	private static void registerSmeltingRecipe(ItemStack input, ItemStack output, float xp) {
 		GameRegistry.addSmelting(input, output, xp);
 	}
 
-	/**
-	 * Registers all smelting recipes
-	 */
 	private static void registerSmeltingRecipes() {
 		float xp = 0.5f;
 		float xpDiamondium = 1.5f;